third_party/blink/web_tests/platform/mac/external/wpt/webmessaging/with-ports/020-expected.txt

This is a testharness.js-based test.
FAIL cross-origin test Uncaught Error: assert_equals: expected 1 but got 2
Harness: the test ran to completion.
